backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
